SIMKDOWN,TGC: Fix compiler error on RPi
On 32-bit Raspbian, "long int" and "long long int" are different, so a
call like

rnd.next(0L, static_cast<int64_t>(blah))

is ambiguous over all overloads of random_t::next

Change the "0L" to static_cast<int64_t>(0) to disambiguate.

+++ my-own-problems/simple-markdown-editor/testcase-generator/simple-markdown-editor-testcase-generator.cpp
@@ -563,7 +563,7 @@ void addRandomishQueries(QueryGenUtils& testcaseGenUtils, const int numQueriesTo
             const auto numCharsToAddToReachTargetLength = targetDocumentLength - testcaseGenUtils.formattingCharsTree.documentLength();
             cout << "numInsertionQueriesToAdd: " << numInsertionQueriesToAdd << " targetDocumentLength: " << targetDocumentLength << " numCharsToAddToReachTargetLength: " << numCharsToAddToReachTargetLength << endl;
             assert(numCharsToAddToReachTargetLength >= numInsertionQueriesToAdd);
-            auto numNonFormattedInsertions = rnd.next(0L, numInsertionQueriesToAdd);
+            auto numNonFormattedInsertions = rnd.next(static_cast<int64_t>(0), numInsertionQueriesToAdd);
             if (!testcaseGenUtils.canRangeQuery() && numNonFormattedInsertions == 0)
                 numNonFormattedInsertions++;
             const auto numFormattedInsertions = numInsertionQueriesToAdd - numNonFormattedInsertions;
@@ -658,7 +658,7 @@ void addRandomishQueries(QueryGenUtils& testcaseGenUtils, const int numQueriesTo
     while (static_cast<int>(testcaseGenUtils.queries.size()) < numQueriesToAdd)
     {
         if (!testcaseGenUtils.canRangeQuery())
-            testcaseGenUtils.addInsertNonFormattingCharQuery(rnd.next(1L, maxDocLength - testcaseGenUtils.formattingCharsTree.documentLength()));
+            testcaseGenUtils.addInsertNonFormattingCharQuery(rnd.next(static_cast<int64_t>(1), maxDocLength - testcaseGenUtils.formattingCharsTree.documentLength()));
         testcaseGenUtils.addIsRangeFormattedQueryBiasingTowardsAfterInsertionPos();
     }
 }
